requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
anthropic==0.5.0
crewai>=0.70.1
litellm>=1.0.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
azure-search-documents==11.4.0
azure-storage-blob==12.17.0
qdrant-client==1.6.0
sentence-transformers==2.2.2
gradio==3.50.0
tiktoken==0.5.1
PyPDF2==3.0.1
fitz==0.0.1.dev2
pymupdf==1.23.4
pandas==2.1.1
openpyxl==3.1.2
python-multipart==0.0.6
aiohttp==3.8.5
beautifulsoup4==4.12.2
pywin32==306; platform_system == "Windows"
docx==0.2.4
comtypes==1.2.0; platform_system == "Windows"
httpx==0.25.0